Creating project.c
